Implemented wrapper classes and delete,update expenditure operations
