--- src/main/java/com/geekydroid/domain/enums/DateFormat.java
@@ -1,5 +1,0 @@
-package com.geekydroid.domain.enums;
-
-public enum DateFormat {
-    POSTGRES_FORMAT
-}
