--- src/main/java/com/geekydroid/domain/enums/Paymode.java
@@ -1,7 +1,0 @@
-package com.geekydroid.domain.enums;
-
-public enum Paymode {
-    CASH,
-    NEFT,
-    UPI
-}
