--- src/main/java/com/geekydroid/domain/enums/TradeType.java
@@ -1,6 +1,0 @@
-package com.geekydroid.domain.enums;
-
-public enum TradeType {
-    BUY,
-    SELL
-}
