--- src/main/java/com/geekydroid/repository/expenditure/ExpenditureRepository.java
@@ -1,17 +1,0 @@
-package com.geekydroid.repository.expenditure;
-
-import com.geekydroid.domain.expenditure.Expenditure;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.transaction.Transactional;
-
-@Transactional
-@ApplicationScoped
-public interface ExpenditureRepository {
-
-    Expenditure createExpenditure(Expenditure expenditure);
-
-    void updateExpenditure(Expenditure expenditure);
-
-    void deleteExpenditure(Expenditure expenditure);
-}
