+++ src/main/java/com/geekydroid/savestmentbackend/domain/enums/DateFormat.java
@@ -1,0 +1,5 @@
+package com.geekydroid.savestmentbackend.domain.enums;
+
+public enum DateFormat {
+    POSTGRES_FORMAT
+}
