+++ src/main/java/com/geekydroid/savestmentbackend/domain/enums/Paymode.java
@@ -1,0 +1,7 @@
+package com.geekydroid.savestmentbackend.domain.enums;
+
+public enum Paymode {
+    CASH,
+    NEFT,
+    UPI
+}
