+++ src/main/java/com/geekydroid/savestmentbackend/domain/enums/TradeType.java
@@ -1,0 +1,6 @@
+package com.geekydroid.savestmentbackend.domain.enums;
+
+public enum TradeType {
+    BUY,
+    SELL
+}
