+++ src/main/java/com/geekydroid/savestmentbackend/repository/expenditure/ExpenditureCategoryRepository.java
@@ -1,0 +1,18 @@
+package com.geekydroid.savestmentbackend.repository.expenditure;
+
+import com.geekydroid.savestmentbackend.domain.expenditure.ExpenditureCategory;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.transaction.Transactional;
+import java.util.List;
+
+@ApplicationScoped
+@Transactional
+public interface ExpenditureCategoryRepository {
+
+    public void createNewExpenditureCategory(ExpenditureCategory expenditureCategory);
+
+    public List<ExpenditureCategory> getAllExpenditureCategories();
+
+    ExpenditureCategory getExpenditureCategoryName(String expenditureCategoryStr);
+}
