+++ src/main/java/com/geekydroid/savestmentbackend/service/expenditure/ExpenditureCategoryService.java
@@ -1,0 +1,16 @@
+package com.geekydroid.savestmentbackend.service.expenditure;
+
+import com.geekydroid.savestmentbackend.domain.expenditure.ExpenditureCategory;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.transaction.Transactional;
+import java.util.List;
+
+@ApplicationScoped
+@Transactional
+public interface ExpenditureCategoryService {
+
+    List<ExpenditureCategory> getAllExpenditureCategories();
+
+    ExpenditureCategory getExpenditureCategoryByName(String expenditureCategoryStr);
+}
