--- src/main/java/com/geekydroid/service/expenditure/ExpenditureCategoryService.java
@@ -1,16 +1,0 @@
-package com.geekydroid.service.expenditure;
-
-import com.geekydroid.domain.expenditure.ExpenditureCategory;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.transaction.Transactional;
-import java.util.List;
-
-@ApplicationScoped
-@Transactional
-public interface ExpenditureCategoryService {
-
-    List<ExpenditureCategory> getAllExpenditureCategories();
-
-    ExpenditureCategory getExpenditureCategoryByName(String expenditureCategoryStr);
-}
