--- src/main/java/com/geekydroid/service/expenditure/ExpenditureCategoryServiceImpl.java
@@ -1,27 +1,0 @@
-package com.geekydroid.service.expenditure;
-
-import com.geekydroid.domain.expenditure.ExpenditureCategory;
-import com.geekydroid.repository.expenditure.ExpenditureCategoryRepository;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.transaction.Transactional;
-import java.util.List;
-
-@ApplicationScoped
-@Transactional
-public class ExpenditureCategoryServiceImpl implements ExpenditureCategoryService{
-
-    @Inject
-    ExpenditureCategoryRepository repository;
-
-    @Override
-    public List<ExpenditureCategory> getAllExpenditureCategories() {
-        return repository.getAllExpenditureCategories();
-    }
-
-    @Override
-    public ExpenditureCategory getExpenditureCategoryByName(String expenditureCategoryStr) {
-        return repository.getExpenditureCategoryName(expenditureCategoryStr);
-    }
-}
